# backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taska API
  description: |
    Taska is a real-time tasking platform connecting customers with local
    taskers for delivery and errand services.

    ## Authentication

    Most endpoints require a JWT access token obtained via the
    `POST /api/v1/auth/verify-otp` flow.

    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```

    The token contains the user's `userId`, `phoneNumber`, and `role` (CUSTOMER | TASKER | ADMIN).
  version: 1.0.0
  contact:
    name: Taska Support
    url: https://taska.app
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /payments:
    post:
      tags: [Payments]
      summary: Create a payment
      description: |
        Creates a new payment for a completed task.
        Supports idempotency via `idempotencyKey`.
        Uses a pluggable payment provider (Stripe, Chapa, Telebirr).
      operationId: createPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentInput'
            example:
              taskId: 550e8400-e29b-41d4-a716-446655440010
              amount: 150.00
              currency: ETB
              description: Payment for grocery delivery
              idempotencyKey: idemp-12345678
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Validation error
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found
          $ref: '#/components/responses/NotFound'
        '409':
          description: Payment already exists for this task
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Payments]
      summary: List payments
      description: |
        Returns payments for the authenticated user.
        Admins can query all payments via `/payments/admin/all`.
      operationId: listPayments
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, PROCESSING, AUTHORIZED, PAID, FAILED, CANCELLED, REFUNDED, PARTIALLY_REFUNDED]
          description: Filter by payment status
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  '/payments/{id}':
    get:
      tags: [Payments]
      summary: Get payment by ID
      description: Returns a single payment with refund history.
      operationId: getPayment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Payment ID
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  '/payments/{id}/confirm':
    post:
      tags: [Payments]
      summary: Confirm payment
      description: |
        Marks a payment as PAID and credits the tasker's wallet.
      operationId: confirmPayment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Payment ID
      responses:
        '200':
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  '/payments/{id}/cancel':
    post:
      tags: [Payments]
      summary: Cancel payment
      description: |
        Cancels a payment that is in PENDING or PROCESSING status.
      operationId: cancelPayment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Payment ID
      responses:
        '200':
          description: Payment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Cannot cancel payment in current status
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  '/payments/{id}/refund':
    post:
      tags: [Payments]
      summary: Process refund
      description: |
        Processes a full or partial refund for a completed payment.
        Debits the tasker's wallet by the refunded amount.
      operationId: refundPayment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Payment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundInput'
            example:
              amount: 50.00
              reason: CUSTOMER_REQUEST
              reasonDetail: Customer reported item damaged
      responses:
        '200':
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
        '400':
          description: Invalid refund amount or status
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  '/payments/{id}/audit-logs':
    get:
      tags: [Payments]
      summary: Get payment audit logs
      description: |
        Returns the append-only audit trail for a payment.
      operationId: getPaymentAuditLogs
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Payment ID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentAuditLog'
        '401':
          $ref: '#/components/responses/Unauthorized'
  '/payments/task/{taskId}':
    get:
      tags: [Payments]
      summary: Get payment by task ID
      description: Returns the payment associated with a specific task.
      operationId: getPaymentByTask
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  '/payments/admin/all':
    get:
      tags: [Payments]
      summary: List all payments (Admin)
      description: |
        Admin-only endpoint. Returns all payments in the system,
        optionally filtered by status.
      operationId: listAllPayments
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, PROCESSING, AUTHORIZED, PAID, FAILED, CANCELLED, REFUNDED, PARTIALLY_REFUNDED]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: All payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Admin access required
          $ref: '#/components/responses/BadRequest'

  # ─────────────────────────────────────────────────────
  #  Wallet
  # ─────────────────────────────────────────────────────
...
  /webhooks/stripe:
    post:
      tags: [Webhooks]
      summary: Stripe webhook
      description: |
        Receives Stripe payment events (payment_intent.succeeded,
        payment_intent.payment_failed). Signature is verified using
        the Stripe webhook secret. No authentication required.
      operationId: handleStripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  example: payment_intent.succeeded
                data:
                  type: object
                  properties:
                    object:
                      type: object
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  received:
                    type: boolean
        '401':
          description: Invalid signature
          $ref: '#/components/responses/Unauthorized'
  /webhooks/generic:
    post:
      tags: [Webhooks]
      summary: Generic webhook
      description: |
        Receives generic payment provider events.
        Expects `event` and `data` fields in the JSON body.
        No authentication required.
      operationId: handleGenericWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event]
              properties:
                event:
                  type: string
                  example: payment_intent.succeeded
                data:
                  type: object
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  received:
                    type: boolean
        '400':
          description: Event type is required
          $ref: '#/components/responses/BadRequest'

  # ─────────────────────────────────────────────────────
  #  Media
  # ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /api/v1/auth/verify-otp`.
        Include as `Authorization: Bearer <token>`.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
      required: [success, error]

    # ── Health ────────────────────────────────────────────
...
    CreatePaymentInput:
      type: object
      required:
        - taskId
        - amount
      properties:
        taskId:
          type: string
          format: uuid
          description: The task being paid for
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Payment amount
        currency:
          type: string
          default: ETB
          description: Currency code (ETB, USD)
        description:
          type: string
          description: Payment description
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          description: Idempotency key to prevent duplicate payments
        metadata:
          type: object
          description: Optional metadata for the payment provider
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        platformFee:
          type: number
          format: double
          description: Platform commission fee
        taskerAmount:
          type: number
          format: double
          description: Amount credited to tasker
        paymentMethod:
          type: string
        paymentStatus:
          type: string
          enum: [PENDING, PROCESSING, AUTHORIZED, PAID, FAILED, CANCELLED, REFUNDED, PARTIALLY_REFUNDED]
        provider:
          type: string
          enum: [STRIPE, CHAPA, TELEBIRR, CASH]
        transactionReference:
          type: string
          nullable: true
        refundedAmount:
          type: number
          format: double
          default: 0
        description:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, taskId, customerId, amount, platformFee, taskerAmount, paymentMethod, paymentStatus, provider, refundedAmount, createdAt, updatedAt]
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Payment'
      required: [success, data]
    RefundInput:
      type: object
      required:
        - amount
        - reason
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Amount to refund (partial or full)
        reason:
          type: string
          enum: [DUPLICATE, FRAUD, CUSTOMER_REQUEST, TASK_CANCELLED, SERVICE_ISSUE, OTHER]
          description: Reason for the refund
        reasonDetail:
          type: string
          description: Optional detailed explanation
    RefundResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            paymentId:
              type: string
              format: uuid
            amount:
              type: number
              format: double
            reason:
              type: string
            status:
              type: string
            providerRefundId:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
    PaymentAuditLog:
      type: object
      properties:
        id:
          type: integer
        paymentId:
          type: string
          format: uuid
        event:
          type: string
          description: The audit event type (CREATED, PROCESSING, COMPLETED, FAILED, etc.)
        fromStatus:
          type: string
          nullable: true
        toStatus:
          type: string
          nullable: true
        data:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── Wallet ─────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Authentication required — missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Authentication required
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Invalid folder. Allowed folders: profile-images, task-images, verification-documents, vehicle-images"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Resource not found
...